rename context menu file
