+++ frontend/javascripts/oxalis/view/context_menu.js
